backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pypdf==5.1.0
faiss-cpu==1.8.0.post1
langchain==0.3.7
langchain-openai==0.2.6
langchain-community==0.3.7
